Format exports (#2)
This formats exports according to single-line and multi-line policies.

The single-line policy simply collapses all exports onto one line, and is
indicated by the input file having the exports on a single line. Spaces
are added after the opening paren and before the closing paren. All
exports are separated by a comma and a space.

The multi-line policy is indicated by the input file having an export
list that spans multiple lines. In this case, it is formatted as an
idiomatic Haskell tuple, with one export per line.